+++ src/views/profile/MedicationTile-style.js
@@ -1,0 +1,7 @@
+export const medListDivStyle = {
+	fontSize: "18px"
+}
+
+export const medListStyle = {
+	listStyleType: 'none'
+}
