--- src/views/profile/Profile_View-style.js
@@ -1,7 +1,0 @@
-export const medListDivStyle = {
-	fontSize: "18px"
-}
-
-export const medListStyle = {
-	listStyleType: 'none'
-}
