+++ src/views/profile/VitalTile.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
-
-import {searchByCode} from '../utils/general_utils.js';
 import $ from 'jquery'; 
+import { searchByCode } from '../../services/risk_score_utils.js';
 
 class VitalTile extends Component {
 	constructor(props) {
@@ -10,6 +9,7 @@ class VitalTile extends Component {
 			name: "",
 			value: "Loading...",
 			units: "",
+			date: ""
 		};
 	}
 
@@ -20,7 +20,7 @@ class VitalTile extends Component {
 			var testobject = {};
 			testobject[parentComponent.props.code] = [];
 			var result = searchByCode(obs, testobject);
-			var precision = 3;
+			var precision = 0;
 			if (result[parentComponent.props.code][0]['value'] < 1) {
 				precision = 2;
 			}
@@ -30,47 +30,57 @@ class VitalTile extends Component {
 			if (result[parentComponent.props.code][0]['text'] === "Low Density Lipoprotein Cholesterol") {
 				result[parentComponent.props.code][0]['text'] = "LDL Cholesterol";
 			}
+			if (result[parentComponent.props.code][0]['text'] === "Systolic Blood Pressure") {
+				result[parentComponent.props.code][0]['text'] = "Systolic BP";
+			}
+			if (result[parentComponent.props.code][0]['text'] === "Diastolic Blood Pressure") {
+				result[parentComponent.props.code][0]['text'] = "Diastolic BP";
+			}
 			var forSparkline = [];
 			for(var i = 0; i < result[parentComponent.props.code].length; i++) {
 				forSparkline.push({
 					name: result[parentComponent.props.code][i]['date'].toString(),
-					value: result[parentComponent.props.code][i]['value']
+					value: (result[parentComponent.props.code][i]['value'])
 				})
 			}
+			//console.log(result);
 			parentComponent.setState({
 				measurementName: result[parentComponent.props.code][0]['text'],
-				units: result[parentComponent.props.code][0]['unit'],
-				value: result[parentComponent.props.code][0]['value'].toPrecision(precision),
-				data: forSparkline
+				value: result[parentComponent.props.code][0]['value'].toFixed(precision) + " " + result[parentComponent.props.code][0]['unit'],
+				data: forSparkline,
+				date: "As of " + result[parentComponent.props.code][0]['date'].slice(0,10)
 			});
 		});
 	}
 	render() {
+		var link = window.location.href + 'measure/' + this.props.code;
 		return (
 			<div>
-				<svg viewBox="0 0 690 106" version="1.1">
+				<svg width="100%" height="100%" xmlns="http://www.w3.org/2000/svg" xmlnsXlink="http://www.w3.org/1999/xlink" viewBox="0 0 750 180" version="1.1">
 	    			<defs>
 	        			<ellipse id="path-1" cx="49.6001408" cy="49.8750284" rx="49.6001408" ry="49.8750284"></ellipse>
 	    			</defs>
-				    <g id="Patient-Page" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-				        <g id="Desktop-HD" transform="translate(-18.000000, -253.000000)">
-				            <g id="Group-4" transform="translate(18.000000, 253.000000)">
-				                <rect id="Rectangle-5" fillOpacity="0.9" fill="#AECEDA" x="0" y="0" width="690" height="106" rx="7.2"></rect>
-			                	{this.props.children}
-				                <text id="Weight" fontFamily="Helvetica" fontSize="30" fontWeight="normal" fill="#000000">
-				                    <tspan x="110" y="38">{this.state.measurementName}</tspan>
-				                </text>
-				                <text id="150-lbs" fontFamily="HiraKakuPro-W3, Hiragino Kaku Gothic Pro" fontSize="40" fontWeight="300" fill="#000000">
-				                    <tspan x="208" y="83">{this.state.value}</tspan>
-				                    <tspan x="286.84" y="83" fontSize="32"> </tspan>
-				                    <tspan x="297.496" y="83" fontSize="20">{this.state.units}</tspan>
-				                </text>
-				                <foreignObject width = "300px" height = "224px" x = "450px" y="40px">
-			                        
-								</foreignObject>
-				            </g>
-				        </g>
-				    </g>
+					    <g id="Patient-Page" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+					        <g id="Desktop-HD" transform="translate(-18.000000, -253.000000)">
+					            <g id="Group-4" transform="translate(18.000000, 253.000000)">
+					            	<a xlinkHref={link} target="_blank">
+					               		<rect id="Rectangle-5" fillOpacity="0.9" fill="#AECEDA" x="0" y="0" width="750" height="180" rx="10" ></rect>
+					               		{this.props.children}
+				                	</a>
+					                <text id="Weight" fontFamily="Helvetica" fontSize="56" fontWeight="normal" fill="#000000">
+					                    <tspan x="180" y="56">{this.state.measurementName}</tspan>
+					                </text>
+					                <text id="150-lbs" fontFamily="HiraKakuPro-W3, Hiragino Kaku Gothic Pro" fontSize="45" fontWeight="300" fill="#000000">
+					                    <tspan x="208" y="130">{this.state.value}</tspan>
+					                    <tspan x="297.496" y="130" fontSize="20">{this.state.units}</tspan>
+					                    <tspan x= "470" y="170" fontSize="30">{this.state.date}</tspan>
+					                </text>
+					                <foreignObject width = "300px" height = "224px" x = "450px" y="40px">
+				                        
+									</foreignObject>
+					            </g>
+					        </g>
+					    </g>
 				</svg>
 			</div>
 		)
